// src/generated/Disk.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Disk = { "type": "ready", name: string, totalBytes: number, freeBytes: number, 
/**
 * Less than a tenth is free.
 */
low: boolean, } | { "type": "unavailable", message: string, };